% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_bf.R
\name{correct_bf}
\alias{correct_bf}
\title{Corrected aggregate Bayes factors}
\usage{
correct_bf(x, y)
}
\arguments{
\item{x}{\code{data.frame} containing mixed model Bayes factors (\code{bf}), sample
size (\code{n_s}), and standard error (\code{sigma_epsilon}).}

\item{y}{\code{data.frame} containing corresponding aggregate log Bayes factors
(\code{bf}).}
}
\value{
The \code{data.frame} passed as \code{x} with an added column \code{logbf_adj}
and the fitted regression model as attribute \code{correction_model}.
}
\description{
Adds corrected log Bayes factors for aggregate analysis based on a linear
regression model predicting the mixed model log Bayes factor from the
aggregate log Bayes factor, sample size, and error variance.
}
