% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_bf.R
\name{correct_ttest_bf}
\alias{correct_ttest_bf}
\title{Corrected aggregate Bayes factors}
\usage{
correct_ttest_bf(x, n_s, sigma_theta, sigma_epsilon)
}
\arguments{
\item{x}{\code{numeric}. Bayes factor from a default paired t-test.}

\item{n_s}{\code{numeric}. Number of subjects.}

\item{sigma_theta}{\code{numeric}. True random slope standard
deviation in the full hierarchical data-generating process.}

\item{sigma_epsilon}{\code{numeric}. True residual standard
deviation in the full hierarchical data-generating process.}
}
\value{
The \code{data.frame} with passed values (\code{x} is stored in
the terribly named column \code{bf2}), the corrected Bayes factor in
the column \code{logbf_adj} and the fitted regression model used
to perform the correction as attribute \code{correction_model}.
}
\description{
Adds corrected log Bayes factors for aggregate analysis based on a linear
regression model predicting the mixed model log Bayes factor from the
aggregate log Bayes factor, sample size, and error variance.
}
\seealso{
\code{\link[BayesFactor:ttestBF]{BayesFactor::ttestBF()}}
}
