% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_translation.R
\name{fixed_effect_rscale}
\alias{fixed_effect_rscale}
\alias{random_effect_rscale}
\alias{ttest_rscale}
\alias{anova_rscale}
\title{Adjust default prior scale for aggregate analysis}
\usage{
fixed_effect_rscale(x, n_t, sigma_theta, sigma_epsilon)

random_effect_rscale(x, n_t, sigma_theta, sigma_epsilon)

ttest_rscale(x, ...)

anova_rscale(x, y, ...)
}
\arguments{
\item{x}{\code{numeric}. Default prior scale r.}

\item{n_t}{\code{numeric}. Number of aggregated trials.}

\item{sigma_theta}{\code{numeric}. True random slope standard
deviation in the full hierarchical data-generating process.}

\item{sigma_epsilon}{\code{numeric}. True residual standard
deviation in the full hierarchical data-generating process.}

\item{...}{Parameters passed to \code{fixed_effect_rscale()} or
\code{random_effect_rscale()}.}
}
\value{
A \code{numeric} of lenght one, which is the adjusted r scale.
\code{anova_rscale()} returns an analogous list with elements
\code{rscaleFixed} and \code{rscaleRandom}.
}
\description{
Adjust the default prior scale from a hierarchical linear mixed
model to yield comparable results in aggregate analyes, i.e.
repeated-measures ANOVA or paired t-test.
}
\examples{
fixed_effect_rscale(0.5, 50, 0.5, 1)
}
\seealso{
\code{\link[BayesFactor:lmBF]{BayesFactor::lmBF()}}, \code{\link[BayesFactor:anovaBF]{BayesFactor::anovaBF()}}, \code{\link[BayesFactor:ttestBF]{BayesFactor::ttestBF()}}, \code{\link[=correct_ttest_bf]{correct_ttest_bf()}}
}
