% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbf_plot.R
\name{logbf_plot}
\alias{logbf_plot}
\title{Scatterplot grid of log Bayes factors}
\usage{
logbf_plot(x, xlab, ylab)
}
\arguments{
\item{x}{\code{data.frame} containing mixed model Bayes factors (\code{bf}), the
corresponding estimation error (\code{error}), effect size (\code{nu}), sample
size (\code{n_s}), number of trials (\code{n_t}), and standard error
(\code{sigma_epsilon}), as well as aggregate Bayes factor (\code{bf2}).}

\item{xlab}{\code{character} or \code{expression} to be used as x-axis label.}

\item{ylab}{\code{character} or `expression to be used as y-axis label.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots mixed model log Bayes factor against the aggregate log Bayes
factors split by all factors varied in the simulation.
}
